pyproject.toml
# pyproject.toml

[tool.poetry]
name = "url-fairy-bot"
version = "2.0.0"
description = "I help to watch content from TikTok, Shorts, Twitter and Instagram Reels more convenient inside Telegram chatgroups"
authors = ["Vladimir nett00n Budylnikov <git@nett00n.org>"]
license = "GPL v3"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
aiogram-utils = "^0.1.4"
fastapi = "^0.111.0"
python = "^3.11"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
yt-dlp = "^2024.5.27"
pydantic-settings = "^2.6.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"
pytest =  "^8.2.0"
yamllint = "^1.35.1"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"

[tool.poetry.scripts]
test = "pytest"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
